拆分统计页面
